% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{By-State Accident Plots}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value specifying the number state to obtain accident data for.}

\item{year}{A list of numeric strings containing the years to obtain accident data for.}
}
\value{
Plot of the state showing the location of accidents in the state. This function will return an error if the year specified does not correspond to a dataset or if the state number does not exist.
}
\description{
This function provides a visual of the number of accidents within a specified state and year.
}
\examples{
fars_map_state(4, 2014)

}
