% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the name of file to be read in.}
}
\value{
This function returns a dataframe table
}
\description{
Checks that the Fatality Analysis Reporting System datafile from the US National Highway
Traffic Safety Administration exists. If it does exist, the data will be read in.
}
\note{
This function will return an error if data is not in the working directory.
}
\examples{
fars_read("accident_2014.csv.bz2")

}
