% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Month/Year Subset}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of numeric strings containing the years to obtain accident data for.}
}
\value{
Returns a list with the months and selected year(s). A separate list is returned for each year. This function will return an error if the specified year does not correspond to a dataset.
}
\description{
This function subsets the data to look only at month and year.
}
\examples{
\dontrun{fars_read_years(2014)}

}
