% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Total Accidents by Month/Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of numeric strings containing the years to obtain summary data for.}
}
\value{
Returns summary as a dataframe. This function will return an error if the year specified does not correspond to a dataset.
}
\description{
Creates by-year summary of the data showing the number of accidents in each month by year.
}
\examples{
fars_summarize_years(c(2014,2015))

}
