% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Define filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric string containing the year to be used to define file name.}
}
\value{
This function returns a character string containing the name of the accident datafile. Example: "accident_2013.csv.bz2". This function will return an error if non-numeric characters are input
}
\description{
Defines name of accident data file (as a csv) based on year provided to function.
}
\note{
This function does not check that the input year is an actual year.
}
\examples{
make_filename(2014)

}
